+++ src/libslic3r/CMakeLists.txt
@@ -116,6 +116,12 @@ add_library(libslic3r STATIC
     Geometry.hpp
     Geometry/Circle.cpp
     Geometry/Circle.hpp
+    Geometry/MedialAxis.Cpp
+    Geometry/MedialAxis.hpp
+    Geometry/Voronoi.hpp
+    Geometry/VoronoiOffset.cpp
+    Geometry/VoronoiOffset.hpp
+    Geometry/VoronoiVisualUtils.hpp
     Int128.hpp
     KDTreeIndirect.hpp
     Layer.cpp
@@ -220,9 +226,6 @@ add_library(libslic3r STATIC
     TriangleSelector.cpp
     TriangleSelector.hpp
     MTUtils.hpp
-    VoronoiOffset.cpp
-    VoronoiOffset.hpp
-    VoronoiVisualUtils.hpp
     Zipper.hpp
     Zipper.cpp
     MinAreaBoundingBox.hpp
