+++ src/libslic3r/Geometry/VoronoiOffset.cpp
@@ -1,5 +1,6 @@
 // Polygon offsetting using Voronoi diagram prodiced by boost::polygon.
 
+#include "Geometry.hpp"
 #include "VoronoiOffset.hpp"
 #include "libslic3r.h"
