+++ tests/libslic3r/test_voronoi.cpp
@@ -6,8 +6,8 @@
 #include <libslic3r/EdgeGrid.hpp>
 #include <libslic3r/Geometry.hpp>
 
-#include <libslic3r/VoronoiOffset.hpp>
-#include <libslic3r/VoronoiVisualUtils.hpp>
+#include <libslic3r/Geometry/VoronoiOffset.hpp>
+#include <libslic3r/Geometry/VoronoiVisualUtils.hpp>
 
 #include <numeric>
